Now the robot says it is your turn only if you didn-t already do a valid move.

+++ src/tictactoe_brain/tictactoeBrain.h
@@ -182,7 +182,7 @@ public:
      * of opponent's tokens in the board increases.
      * @param number of opponent's token at the beginning
      **/
-    void waitForOpponentTurn(const uint8_t& n_robot_tokens);
+    void waitForOpponentTurn(const size_t& n_robot_tokens);
 
     /**
      * Indicates if the board is full.
